cleanup

+++ cute_kernels/kernels/cross_entropy/torch_implementation.py
@@ -3,4 +3,4 @@ import torch.nn.functional as F
 
 
 def cross_entropy_torch(x: torch.Tensor, labels: torch.Tensor, reduction: str = "mean") -> torch.Tensor:
-    return F.cross_entropy(x, labels, reduction=reduction)
+    return F.cross_entropy(x.float(), labels, reduction=reduction)
